Questão 1, 2 e makefile
Makefile aletrado para compilar as 3 questões
Adicionados logs do gprof da questão 2
Questão 1

+++ include/questao_02/fatorial.h
@@ -2,6 +2,6 @@
 #define FATORIAL_H
 
 // Calculo do Fatorial
-int fat(int num);
+long int fat(int num);
 
 #endif
